application/requirements.txt
Flask==0.12
requests==2.18.4
flask-restful==0.3.5
flasgger==0.6.6
Flask-CORS===3.0.3
Flask-PyMongo
Gunicorn==19.8.1
gevent==1.3.4
